Adds test for inclusion of vendor paths
This provides a test that checks if the moq tool is able to handle
vendored imports. This is currently a problem when used with the
standard `go generate` tool, which this test is able to show.

The test requires you to "go install" moq, but should work with any
version of Go. Reproducibility is there an issue if these conditions are
not met.